pyproject.toml
[project]
name = "Wavelet-Transformer-Downsampler"
version = "0.1.0"
description = "a Python package for time series downsampling using wavelet transforms and transformer models"
readme = "README.md"
authors = [{name = "Ehsan Honarbakhsh", email = "ehsan.honarbakhsh96@gmail.com"}]
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.2.2",
    "astunparse==1.6.3",
    "attrs==25.3.0",
    "avro==1.12.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "cloudpickle==2.2.1",
    "contourpy==1.3.2",
    "crcmod==1.7",
    "cycler==0.12.1",
    "dill==0.3.1.1",
    "dnspython==2.7.0",
    "docopt==0.6.2",
    "fastavro==1.10.0",
    "fasteners==0.19",
    "find_libpython==0.4.0",
    "flatbuffers==25.2.10",
    "fonttools==4.57.0",
    "gast==0.6.0",
    "google-pasta==0.2.0",
    "grpcio==1.71.0",
    "h5py==3.13.0",
    "hdfs==2.7.3",
    "httplib2==0.22.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "joblib==1.5.0",
    "jsonpickle==3.4.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "keras==3.9.2",
    "kiwisolver==1.4.8",
    "libclang==18.1.1",
    "Markdown==3.8",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.3",
    "mdurl==0.1.2",
    "ml-dtypes==0.3.2",
    "namex==0.0.9",
    "numpy==1.26.4",
    "objsize==0.7.1",
    "opt_einsum==3.4.0",
    "optree==0.15.0",
    "orjson==3.10.18",
    "packaging==25.0",
    "pandas==2.2.3",
    "pemja==0.4.1",
    "pillow==11.2.1",
    "pluggy==1.6.0",
    "proto-plus==1.26.1",
    "protobuf==4.25.7",
    "psutil==7.0.0",
    "py4j==0.10.9.7",
    "pyarrow==16.1.0",
    "pyarrow-hotfix==0.7",
    "pydot==1.4.2",
    "Pygments==2.19.1",
    "PyJWT==2.9.0",
    "pymongo==4.12.1",
    "pyparsing==3.2.3",
    "pytest==8.4.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyWavelets==1.8.0",
    "PyYAML==6.0.2",
    "redis==5.3.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.0.0",
    "rpds-py==0.24.0",
    "ruamel.yaml==0.18.10",
    "ruamel.yaml.clib==0.2.12",
    "scikit-learn==1.6.1",
    "scipy==1.15.3",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "tensorboard==2.16.2",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.16.2",
    "tensorflow-addons==0.23.0",
    "tensorflow-io-gcs-filesystem==0.37.1",
    "termcolor==3.1.0",
    "threadpoolctl==3.6.0",
    "typeguard==2.13.3",
    "typing_extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "Werkzeug==3.1.3",
    "wrapt==1.17.2",
    "zstandard==0.23.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/ehsan-honarbakhsh/Wavelet-Transformer-Downsampler"
Repository = "https://github.com/ehsan-honarbakhsh/Wavelet-Transformer-Downsampler"

[tool.hatch.build.targets.wheel]
packages = ["wavelet_transformer_downsampler"]